+++ CPP/src/main.cpp
@@ -20,6 +20,27 @@ int main(){
     masterCardList.find(i)->second.printCard();
     cout << endl << endl;
   }
+
+  map<string, Profile> profileDB;
+  loadProfileData(tempFileString, &profileDB, &masterCardList, true);
+
+  Profile tmpProfile = profileDB.at("Funshine");
+  vector<Deck> tmpDeckList = tmpProfile.getAllDecks();
+
+  for (uint8_t i; i < tmpDeckList.size(); i++)
+  {
+    cout << "Deck: " << tmpDeckList.at(i).getDeckName() << endl;
+
+    vector<Card> tmpCardList = tmpDeckList.at(i).getOriginalDeck();
+    for (uint8_t j = 0; j < tmpCardList.size(); j++)
+    {
+      cout << "\t" << (j+1) << "." << tmpCardList.at(j).getCardName();
+      cout << endl << endl;
+    }
+    
+  }
+
+
   cout << "TEST END" << endl;
 
   return 0;